Remove default empty metadata type
