+++ Sources/Player/Types/DirectAssetLoader.swift
@@ -6,23 +6,23 @@
 
 import Combine
 
-enum DirectAssetLoader: AssetLoader {
+enum DirectAssetLoader<CustomData>: AssetLoader {
     struct Input {
         let asset: Asset
-        let metadata: PlayerMetadata
+        let metadata: AssetMetadata<CustomData>
     }
 
-    static func metadataPublisher(for input: Input) -> AnyPublisher<PlayerMetadata, any Error> {
+    static func metadataPublisher(for input: Input) -> AnyPublisher<AssetMetadata<CustomData>, any Error> {
         Just(input.metadata)
             .setFailureType(to: Error.self)
             .eraseToAnyPublisher()
     }
 
-    static func asset(from input: Input, metadata: PlayerMetadata) -> Asset {
+    static func asset(from input: Input, metadata: AssetMetadata<CustomData>) -> Asset {
         input.asset
     }
 
-    static func playerMetadata(from input: Input, metadata: PlayerMetadata?) -> PlayerMetadata {
-        input.metadata
+    static func playerMetadata(from input: Input, metadata: AssetMetadata<CustomData>?) -> PlayerMetadata {
+        input.metadata.playerMetadata
     }
 }
