+++ Sources/StandardConnector/Extensions/PlayerItem.swift
@@ -34,25 +34,4 @@ public extension PlayerItem {
             trackerAdapters: trackerAdapters
         )
     }
-
-    /// Creates a player item from a backend endpoint providing Pillarbox-standard metadata.
-    ///
-    /// - Parameters:
-    ///   - request: The `URLRequest` to fetch metadata from.
-    ///   - decoder: A `JSONDecoder` used to decode the response. Defaults to `JSONDecoder()`.
-    ///   - trackerAdapters: An array of `TrackerAdapter` instances to use for tracking playback events.
-    ///   - assetProvider: A closure that receives the decoded `PlayerData` and returns an `Asset`.
-    ///
-    /// This method fetches metadata from the provided endpoint and decodes it into a `PlayerData` object.
-    ///
-    /// The connector does not create the `Asset` automatically. Users must provide an `assetProvider` closure that transforms
-    /// the decoded `PlayerData` into a playable `Asset`.
-    static func standard(
-        request: URLRequest,
-        decoder: JSONDecoder = JSONDecoder(),
-        trackerAdapters: [TrackerAdapter<PlayerData<EmptyCustomData>>] = [],
-        assetProvider: @escaping (PlayerData<EmptyCustomData>) -> Asset
-    ) -> Self {
-        Self.standard(request: request, type: EmptyCustomData.self, decoder: decoder, trackerAdapters: trackerAdapters, assetProvider: assetProvider)
-    }
 }
