+++ Sources/StandardConnector/PlayerData.swift
@@ -8,10 +8,6 @@ import CoreMedia
 import Foundation
 import PillarboxPlayer
 
-// FIXME: Remove when minimum target is 17 and use `Never` instead.
-/// An object representing the absence of custom data.
-public struct EmptyCustomData: Decodable {}
-
 /// Metadata associated with content loaded in a player.
 ///
 /// Represents the standard metadata returned by a backend endpoint and used to configure a playable `Asset`.
